+++ lib/db_core/models/db_state.dart
@@ -3,7 +3,16 @@ import 'db_errors.dart';
 
 part 'db_state.freezed.dart';
 
-enum DatabaseStatus { idle, open, locked, loading, closingSync, closed, error }
+enum DatabaseStatus {
+  idle,
+  opening,
+  open,
+  locked,
+  loading,
+  closingSync,
+  closed,
+  error,
+}
 
 @freezed
 @immutable
@@ -19,6 +28,7 @@ sealed class DatabaseState with _$DatabaseState {
   const DatabaseState._();
 
   bool get isIdle => status == DatabaseStatus.idle;
+  bool get isOpening => status == DatabaseStatus.opening;
   bool get isOpen => status == DatabaseStatus.open;
   bool get isClosed => status == DatabaseStatus.closed;
   bool get isLocked => status == DatabaseStatus.locked;